requirements.txt
connexion>=1.0.109
tox>=2.3.1
Flask-PyMongo>=0.4.1
swagger-parser>=0.1.10